# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Kestrel AI backend API
  version: 1.0.0
  description: |
    Contract between the Tauri/React frontend and the FastAPI sidecar. This file is the
    source of truth (spec section 2). The backend validates responses against it in tests,
    the TypeScript client is generated from it, and Prism serves it as the mock server.

    Conventions:
    - Every path below carries the `/api/v1` prefix. JSON bodies, UTF-8.
    - Every request carries the per-launch token, either as `Authorization: Bearer <token>`
      or as a `token` query parameter (needed for `<img>` sources and websocket connects).
    - Ids are UUID strings. Timestamps are RFC 3339 strings in UTC.
    - Boxes are in image pixels: `x, y` top-left corner, `w, h` size.
    - Every list endpoint paginates with `limit` and `cursor` and returns `next_cursor`
      (`null` when there is no further page).
    - Every error uses the `Error` envelope `{error: {code, message, details}}`.
    - Long work returns a `Job`; progress arrives over the websocket described in
      `x-websocket` and can be polled through the `jobs` resource.
servers:
  - url: http://127.0.0.1:8765
    description: development backend origin (the packaged app uses a random free port)
  - url: http://127.0.0.1:4010
    description: Prism mock server
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [projects]
      operationId: getProject
      responses:
        "200":
          description: the project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      operationId: updateProject
      summary: Update name, pre-annotation model or import defaults (project settings screen).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      operationId: forgetProject
      summary: Remove the project from the recent list. The project folder and everything in it stay on disk; "Open folder" brings it back.
      responses:
        "204":
          description: forgotten
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/projects/{projectId}/site-areas:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [detect]
      operationId: listSiteAreas
      responses:
        "200":
          description: site areas, oldest first
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteAreaList" }
              example:
                items:
                  - id: "5a000000-aaaa-4000-8000-000000000001"
                    name: "North laydown yard"
                    polygon_wgs84: [[47.7610, 29.4960], [47.7640, 29.4960], [47.7640, 29.4980], [47.7610, 29.4980]]
                    created_at: "2026-09-23T09:00:00Z"
                  - id: "5a000000-aaaa-4000-8000-000000000002"
                    name: "Batching plant"
                    polygon_wgs84: [[47.7650, 29.4930], [47.7680, 29.4930], [47.7680, 29.4950]]
                    created_at: "2026-09-23T09:05:00Z"
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
    post:
      tags: [detect]
      operationId: createSiteArea
      summary: |
        Add a site area, given in WGS84 or as a polygon drawn in one map's pixels (converted with
        that map's georeference). Starts an `area_recount` job over every map run.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SiteAreaCreate" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteArea" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/site-areas/{areaId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/areaId"
    patch:
      tags: [detect]
      operationId: updateSiteArea
      summary: Rename or redraw a site area. A new polygon starts an `area_recount` job.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SiteAreaPatch" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SiteArea" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [detect]
      operationId: deleteSiteArea
      summary: Delete a site area. Starts an `area_recount` job.
      responses:
        "204": { description: deleted }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/analytics/sources/{sourceId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/sourceId"
    get:
      tags: [detect]
      operationId: getSourceAnalytics
      summary: Counts per class for one source from its chosen run (pinned, else the comparison rule for maps and the newest run for photos). Reads run rows only.
      responses:
        "200":
          description: the source's counts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceAnalytics" }
              example:
                source:
                  id: "50000000-3333-4000-8000-000000000002"
                  kind: map
                  label: "May survey"
                  captured_on: "2026-05-20"
                  map_id: "a0000000-6666-4000-8000-000000000001"
                  folder: "D:\\Surveys\\may-ortho.tif"
                  site: may-ortho
                  settings: { max_side: 4000, quality: 95, dedupe_threshold: 4 }
                  image_count: 0
                  duplicate_count: 0
                  job_id: "j0000000-4444-4000-8000-000000000002"
                  imported_at: "2026-05-21T10:00:00Z"
                  created_at: "2026-05-21T09:55:00Z"
                unit: objects
                image_count: null
                run:
                  id: "r0000000-7777-4000-8000-000000000001"
                  kind: map
                  source_id: "50000000-3333-4000-8000-000000000002"
                  source_label: "May survey"
                  model_id: "m0000000-2222-4000-8000-000000000001"
                  model_name: "machinery-v3"
                  conf: 0.25
                  job_state: succeeded
                  pinned: false
                  counts: { "c1a2b3c4-0000-4000-8000-000000000001": 42, "c1a2b3c4-0000-4000-8000-000000000004": 17 }
                  verified_counts: { "c1a2b3c4-0000-4000-8000-000000000001": 30 }
                  review: { total: 59, reviewed: 34 }
                  created_at: "2026-09-22T11:00:00Z"
                classes:
                  - { class_id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", total: 42, verified: 30 }
                  - { class_id: "c1a2b3c4-0000-4000-8000-000000000004", name: "dump truck", colour: "#0ea5e9", total: 17, verified: 0 }
                review: { total: 59, reviewed: 34 }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/analytics/areas:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [detect]
      operationId: getAreaAnalytics
      summary: Counts per site area and class for every survey, in survey-timeline order and on the same comparison basis. Reads run rows only.
      parameters:
        - { name: model_id, in: query, required: false, schema: { type: string }, description: compare on this model instead of the newest run's }
        - { name: conf, in: query, required: false, schema: { type: number, minimum: 0, maximum: 1 }, description: compare at this confidence instead of the newest run's }
        - { name: verified_only, in: query, required: false, schema: { type: boolean, default: false }, description: pick surveys as the verified-only timeline does }
      responses:
        "200":
          description: per-area counts
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AreaAnalytics" }
              example:
                areas:
                  - { id: "5a000000-aaaa-4000-8000-000000000001", name: "North laydown yard" }
                  - { id: "5a000000-aaaa-4000-8000-000000000002", name: "Batching plant" }
                surveys:
                  - map_id: "7c9e1b2a-5555-4000-8000-000000000001"
                    map_name: "April survey"
                    captured_on: "2026-04-15"
                    state: ok
                    per_area:
                      "5a000000-aaaa-4000-8000-000000000001":
                        partial: false
                        counts: { "c1a2b3c4-0000-4000-8000-000000000001": { total: 5, verified: 5 }, "c1a2b3c4-0000-4000-8000-000000000004": { total: 2, verified: 1 } }
                      "5a000000-aaaa-4000-8000-000000000002":
                        partial: true
                        counts: { "c1a2b3c4-0000-4000-8000-000000000001": { total: 1, verified: 0 } }
                  - map_id: "7c9e1b2a-5555-4000-8000-000000000002"
                    map_name: "May survey"
                    captured_on: "2026-05-20"
                    state: ok
                    per_area:
                      "5a000000-aaaa-4000-8000-000000000001":
                        partial: false
                        counts: { "c1a2b3c4-0000-4000-8000-000000000001": { total: 7, verified: 6 }, "c1a2b3c4-0000-4000-8000-000000000004": { total: 3, verified: 3 } }
                      "5a000000-aaaa-4000-8000-000000000002":
                        partial: false
                        counts: { "c1a2b3c4-0000-4000-8000-000000000001": { total: 2, verified: 2 } }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/analytics/photo-batches:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [detect]
      operationId: getPhotoBatchAnalytics
      summary: Detections per class for every photo source. These are detections, not objects - the same object appears in several photos - and never enter a trend.
      responses:
        "200":
          description: one row per photo source
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PhotoBatchAnalytics" }
              example:
                batches:
                  - source:
                      id: "50000000-3333-4000-8000-000000000001"
                      kind: images
                      label: "Flight 15 Apr"
                      captured_on: "2019-04-15"
                      map_id: null
                      folder: "E:\\Dev\\Yolo\\Ahmadia Construction Data"
                      site: ahmadia
                      settings: { max_side: 4000, quality: 95, dedupe_threshold: 4 }
                      image_count: 3299
                      duplicate_count: 0
                      job_id: "j0000000-4444-4000-8000-000000000001"
                      imported_at: "2026-09-17T10:30:00Z"
                      created_at: "2026-09-17T10:05:00Z"
                    run:
                      id: "q0000000-8888-4000-8000-000000000001"
                      kind: images
                      source_id: "50000000-3333-4000-8000-000000000001"
                      source_label: "Flight 15 Apr"
                      model_id: "m0000000-2222-4000-8000-000000000001"
                      model_name: "machinery-v3"
                      conf: 0.25
                      job_state: succeeded
                      pinned: false
                      counts: { "c1a2b3c4-0000-4000-8000-000000000001": 31, "c1a2b3c4-0000-4000-8000-000000000004": 9 }
                      verified_counts: { "c1a2b3c4-0000-4000-8000-000000000001": 12 }
                      review: { total: 40, reviewed: 12 }
                      created_at: "2026-09-21T15:00:00Z"
                    classes:
                      - { class_id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", total: 31, verified: 12 }
                      - { class_id: "c1a2b3c4-0000-4000-8000-000000000004", name: "dump truck", colour: "#0ea5e9", total: 9, verified: 0 }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        default: { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: the per-launch token passed to the sidecar in `APP_TOKEN`
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: same token as a query parameter, for image URLs and the websocket
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
    sourceId:
      name: sourceId
      in: path
      required: true
      schema: { type: string }
...
    areaId:
      name: areaId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    WrongProjectKind:
      description: "the operation does not belong to this kind of project (`code` is `wrong_project_kind`, details `{kind, allowed}`)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: wrong_project_kind
              message: "This is a detection project. Datasets belong in a training project."
              details: { kind: detect, allowed: [train] }
...
    NotFound:
      description: "the project or the named resource does not exist (`code` is `not_found`)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: "run r0000000… not found", details: {} }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: |
                machine-readable: unauthorized, not_found, validation_error, already_exists,
                class_in_use, conflict, not_implemented, provider_error, internal_error,
                wrong_project_kind (409: the operation does not belong to this kind of project;
                details `{kind, allowed}`), library_unavailable (503: the model library could not
                be opened), model_unavailable (409: the library model's weights file is missing),
                unmapped_classes (422: the run's model has classes with no project class and no
                remembered mapping; details `{model_id, unmapped}`), unsupported_point_cloud,
                insufficient_memory and insufficient_disk (422: a point cloud cannot be read or
                admitted), link_needs_coordinates, no_overlap and crs_already_set (422: a point
                cloud patch), invalid_epsg (422: an EPSG code pyproj does not know),
                wrong_point_count, needs_projected_crs, vertical_span_too_small and
                measurement_limit (422: a point-cloud measurement), octree_missing (404: a ready
                cloud's display-copy file is missing on disk), grid_too_large (422: a surface
                grid over the cell ceiling), unsupported_crs, source_missing and
                invalid_build_request (422: a surface build), no_coordinates (422: an ortho
                underlay without a CRS), invalid_geometry and invalid_base (422: a volume
                measurement's polygon or base), job_running (409: the resource's job is queued
                or running), not_ready (409: the resource has not finished importing),
                range_not_satisfiable (416)
            message: { type: string }
            details:
              type: object
              additionalProperties: true
      example:
        error: { code: not_found, message: "project 7f1c… not found", details: {} }
...
    ClassDef:
      type: object
      required: [id, name, colour, hotkey, order]
      properties:
        id: { type: string }
        name: { type: string }
        colour: { type: string }
        hotkey: { type: [string, "null"] }
        order: { type: integer }
      example: { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
    ImportSettings:
      type: object
      description: preparation settings (spec section 5); missing fields take the project defaults
      properties:
        max_side: { type: integer, minimum: 512, maximum: 12000, default: 4000 }
        quality: { type: integer, minimum: 50, maximum: 100, default: 95 }
        dedupe_threshold: { type: integer, minimum: 0, maximum: 32, default: 4, description: max phash Hamming distance for a duplicate }
        group_regex:
          type: string
          default: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)"
          description: the named group `flight` is the group key (flights span cameras); fall back to a 250 m tile, then the source site
      example: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
    ProjectKind:
      type: string
      enum: [train, detect]
      description: "`train` labels images, builds datasets and trains models; `detect` runs library models over sources. Set at creation, never changed."
      example: train
    Project:
      type: object
      required: [id, name, folder, kind, classes, preannotation_model_id, import_defaults, schema_version, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        folder: { type: string, description: absolute path of the project folder }
        kind: { $ref: "#/components/schemas/ProjectKind" }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClassDef" }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
        schema_version: { type: integer }
        created_at: { type: string, format: date-time }
      example:
        id: "7f1c2e3a-1111-4000-8000-000000000001"
        name: Ahmadia
        folder: "E:\\Projects\\Ahmadia"
        kind: train
        classes:
          - { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000002", name: wheel_loader, colour: "#eab308", hotkey: "2", order: 1 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000003", name: bulldozer, colour: "#22c55e", hotkey: "3", order: 2 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000004", name: dump_truck, colour: "#06b6d4", hotkey: "4", order: 3 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000005", name: crane, colour: "#3b82f6", hotkey: "5", order: 4 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000006", name: concrete_mixer, colour: "#a855f7", hotkey: "6", order: 5 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000007", name: roller, colour: "#ec4899", hotkey: "7", order: 6 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000008", name: backhoe, colour: "#ef4444", hotkey: "8", order: 7 }
        preannotation_model_id: "m0000000-2222-4000-8000-000000000001"
        import_defaults: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
        schema_version: 1
        created_at: "2026-09-17T10:00:00Z"
...
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
      example: { name: Ahmadia, preannotation_model_id: "m0000000-2222-4000-8000-000000000001" }
...
    Source:
      type: object
      required: [id, kind, label, captured_on, map_id, folder, site, settings, image_count, duplicate_count, job_id, imported_at, created_at]
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/SourceKind" }
        label: { type: [string, "null"], description: "the operator's name for the source, e.g. Flight 14 Sep; a map source starts with the map's name" }
        captured_on:
          type: [string, "null"]
          format: date
          description: "the survey date. Photos: the earliest capture time in the import, editable. Maps: the map's `captured_on`, kept equal to it. Null means date not set, and it is never guessed."
        map_id: { type: [string, "null"], description: the map a `map` source owns; null for photos }
        folder: { type: string, description: the imported folder; originals there are never modified }
        site: { type: string }
        settings: { $ref: "#/components/schemas/ImportSettings" }
        image_count: { type: integer }
        duplicate_count: { type: integer }
        job_id: { type: [string, "null"], description: the latest import job }
        imported_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
      example:
        id: "50000000-3333-4000-8000-000000000001"
        kind: images
        label: "Flight 15 Apr"
        captured_on: "2019-04-15"
        map_id: null
        folder: "E:\\Dev\\Yolo\\Ahmadia Construction Data"
        site: ahmadia
        settings: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
        image_count: 3299
        duplicate_count: 0
        job_id: "j0000000-4444-4000-8000-000000000001"
        imported_at: "2026-09-17T10:30:00Z"
        created_at: "2026-09-17T10:05:00Z"
...
    LonLat:
      type: array
      description: "[longitude, latitude] in WGS84 degrees"
      minItems: 2
      maxItems: 2
      items: { type: number }
    PixelPoint:
      type: array
      description: "[x, y] in full-resolution map pixels"
      minItems: 2
      maxItems: 2
      items: { type: number }
    SiteArea:
      type: object
      required: [id, name, polygon_wgs84, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        polygon_wgs84:
          type: array
          minItems: 3
          description: the outline, not closed (the first point is not repeated)
          items: { $ref: "#/components/schemas/LonLat" }
        created_at: { type: string, format: date-time }
      example:
        id: "5a000000-aaaa-4000-8000-000000000001"
        name: "North laydown yard"
        polygon_wgs84: [[47.7610, 29.4960], [47.7640, 29.4960], [47.7640, 29.4980], [47.7610, 29.4980]]
        created_at: "2026-09-23T09:00:00Z"
    SiteAreaCreate:
      type: object
      description: "either `polygon_wgs84`, or `map_id` with `polygon_px` drawn on that map (the server converts it with the map's georeference)"
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        polygon_wgs84:
          type: array
          minItems: 3
          items: { $ref: "#/components/schemas/LonLat" }
        map_id: { type: string }
        polygon_px:
          type: array
          minItems: 3
          items: { $ref: "#/components/schemas/PixelPoint" }
      example: { name: "North laydown yard", map_id: "a0000000-6666-4000-8000-000000000001", polygon_px: [[1000, 800], [4000, 800], [4000, 3000], [1000, 3000]] }
    SiteAreaPatch:
      type: object
      description: "a new outline is either `polygon_wgs84`, or `map_id` with `polygon_px`"
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        polygon_wgs84:
          type: array
          minItems: 3
          items: { $ref: "#/components/schemas/LonLat" }
        map_id: { type: string }
        polygon_px:
          type: array
          minItems: 3
          items: { $ref: "#/components/schemas/PixelPoint" }
      example: { name: "North yard" }
    SiteAreaList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/SiteArea" }
    ClassCountRow:
      type: object
      required: [class_id, name, colour, total, verified]
      properties:
        class_id: { type: string }
        name: { type: string }
        colour: { type: string }
        total: { type: integer, minimum: 0 }
        verified: { type: integer, minimum: 0 }
      example: { class_id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", total: 6, verified: 4 }
    SourceAnalytics:
      type: object
      required: [source, unit, image_count, run, classes, review]
      properties:
        source: { $ref: "#/components/schemas/Source" }
        unit:
          type: string
          enum: [objects, detections]
          description: "`objects` for a map source; `detections` for photos, where one object can appear in several photos"
        image_count: { type: [integer, "null"], description: photos in a photo source; null for a map }
        run:
          description: the run that represents the source; null before any run
          oneOf:
            - { $ref: "#/components/schemas/RunSummary" }
            - { type: "null" }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClassCountRow" }
        review: { $ref: "#/components/schemas/ReviewProgress" }
...
    AreaAnalytics:
      type: object
      required: [areas, surveys]
      properties:
        areas:
          type: array
          items: { $ref: "#/components/schemas/AreaRef" }
        surveys:
          type: array
          description: oldest first, as in the survey timeline
          items: { $ref: "#/components/schemas/AreaSurvey" }
...
    PhotoBatchAnalytics:
      type: object
      required: [batches]
      properties:
        batches:
          type: array
          items: { $ref: "#/components/schemas/PhotoBatch" }
...